Your descriptive message

+++ app/api/feed/[id]/analytics/route.ts
@@ -23,8 +23,8 @@ export async function GET(
   }>(
     `select
        s.view_count, s.total_watch_seconds, s.share_count,
-       greatest(0, (select count(*) from video_likes vl where vl.submission_id = s.id and vl.value = 1)::int + s.like_offset) as likes,
-       greatest(0, (select count(*) from video_likes vl where vl.submission_id = s.id and vl.value = -1)::int + s.dislike_offset) as dislikes,
+       greatest(0, (select count(*) from video_likes vl where vl.submission_id = s.id and vl.value = 1) + s.like_offset) as likes,
+       greatest(0, (select count(*) from video_likes vl where vl.submission_id = s.id and vl.value = -1) + s.dislike_offset) as dislikes,
        (select count(*) from video_comments vc where vc.submission_id = s.id and not vc.deleted)::int as comment_count,
        (select count(*) from video_reports vr where vr.submission_id = s.id)::int as report_count,
        (select avg(watch_seconds) from video_views vv where vv.submission_id = s.id) as avg_watch_seconds,
